backend/requirements.txt
Django==5.1
gunicorn==23.0.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
pymysql==1.1.0
PyMySQL==1.1.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
django-environ==0.11.2
drf-spectacular==0.27.2
pillow==11.0.0
boto3==1.35.54
python-decouple==3.8
mysqlclient==2.2.6
pandas==2.2.3
gprofiler-official==1.0.0
django-storages==1.14.4
scipy==1.15.1
matplotlib==3.10.0
django-ses==4.4.0
djangorestframework-simplejwt==5.5.0 
pyjwt==2.9.0